completed code

+++ routes/product-routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const { Product, Category, Tag, ProductTag } = require('../../models')
+const { Product, Category, Tag, ProductTag } = require('../models')
 
 // The `/api/products` endpoint
 
@@ -7,12 +7,26 @@ const { Product, Category, Tag, ProductTag } = require('../../models')
 router.get('/products', (req, res) => {
   // find all products
   // be sure to include its associated Category and Tag data
+  Product.findAll({include: [Category, Tag]})
+  .then(products => {
+    res.json(products)
+  })
+  .catch(err => {
+    console.log(err)
+  })
 })
 
 // get one product
 router.get('/products/:id', (req, res) => {
   // find a single product by its `id`
   // be sure to include its associated Category and Tag data
+  Product.findOne({ where: { id: req.params.id }, include: [Category, Tag] })
+    .then(product => {
+      res.json(product)
+    })
+    .catch(err => {
+      console.log(err)
+    })
 })
 
 // create new product
@@ -22,14 +36,14 @@ router.post('/products', (req, res) => {
       product_name: "Basketball",
       price: 200.00,
       stock: 3,
-      tagIds: [1, 2, 3, 4]
+      category_id: [1, 2, 3, 4]
     }
   */
   Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      if (req.body.tagIds.length) {
-        const productTagIdArr = req.body.tagIds.map((tag_id) => {
+      if (req.body.category_id.length) {
+        const productTagIdArr = req.body.category_id.map((tag_id) => {
           return {
             product_id: product.id,
             tag_id
@@ -52,45 +66,53 @@ router.put('/products/:id', (req, res) => {
   // update product data
   Product.update(req.body, {
     where: {
-      id: req.params.id
-    }
+      id: req.params.id,
+    },
   })
     .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } })
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id)
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id
-          }
-        })
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id)
+      if (req.body.categoryid && req.body.categoryid.length) {
+        const productTags = ProductTag.findAll({
+          where: { product_id: req.params.id }
+        });
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.categoryid
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.categoryid.includes(tag_id))
+          .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags)
-      ])
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      }
+
+      return res.status(200).json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
       // console.log(err);
-      res.status(400).json(err)
-    })
-})
+      res.status(400).json(err);
+    });
+});
 
 router.delete('/products/:id', (req, res) => {
   // delete one product by its `id` value
+  Product.destroy({ where: { id: req.params.id } })
+    .then(product => {
+      res.sendStatus(200)
+    })
+    .catch(err => {
+      console.log(err)
+    })
 })
 
 module.exports = router
